changed User id type to Integer from Long (that's a lot of users) added removeRoleFromUser to SecurityService

+++ devnull-security/src/main/groovy/org/devnull/security/dao/UserDao.groovy
@@ -4,7 +4,7 @@ import org.devnull.security.model.User
 import org.springframework.data.repository.PagingAndSortingRepository
 import org.springframework.data.jpa.repository.Modifying
 
-interface UserDao extends PagingAndSortingRepository<User, Long> {
+interface UserDao extends PagingAndSortingRepository<User, Integer> {
     User findByOpenId(String openId)
 
     @Modifying
